added synchronization so everything should hopefully be thread safe. there should be no way to screw things up by pressing a button at the wrong time.
also fixed a few bugs where the UI was not updating when it should have been.

changed buttons to display icons instead of text.

no longer crashes when trying to move past the ends of the song while it is not playing.
I do need to decide what I want it to do, though. for example, should it move to the other end or just do nothing?
perhaps there should be a setting for this.