trying to fix fragments, but it is still craching in many cases upon rotation.
I think the issue is that the automatic fragment transition backstack restoration does not work well for restoring.
I will need to manually make sure that states are saved and restored during the life cycle of the app or that no states are saved and it reconsructs everything when config changes.