fixed a lot of small issues.
fixed layout on custom number picker.
fixed scrolling to the bottom when adding new pickers to the pattern picker.
buttons to add/remove pickers are now properly enabled/disabled when they should be.
modified custom number picker to work with numbers when the minimum is not 0. tested and it works.